refactor: move core files out of vda5050_core and delete pkg
