+++ include/vda5050_core/transport/paho_mqtt_client.hpp
@@ -16,8 +16,8 @@
  * limitations under the License.
  */
 
-#ifndef VDA5050_CORE__MQTT_CLIENT__PAHO_MQTT_CLIENT_HPP_
-#define VDA5050_CORE__MQTT_CLIENT__PAHO_MQTT_CLIENT_HPP_
+#ifndef VDA5050_CORE__TRANSPORT__PAHO_MQTT_CLIENT_HPP_
+#define VDA5050_CORE__TRANSPORT__PAHO_MQTT_CLIENT_HPP_
 
 #include <mqtt/async_client.h>
 #include <mqtt/callback.h>
@@ -30,11 +30,11 @@
 #include <string>
 #include <unordered_map>
 
-#include "vda5050_core/mqtt_client/mqtt_client_interface.hpp"
+#include "vda5050_core/transport/mqtt_client_interface.hpp"
 
 namespace vda5050_core {
 
-namespace mqtt_client {
+namespace transport {
 
 class PahoMqttClient;
 
@@ -158,7 +158,7 @@ private:
   mqtt::connect_options conn_options_;
 };
 
-}  // namespace mqtt_client
+}  // namespace transport
 }  // namespace vda5050_core
 
-#endif  // VDA5050_CORE__MQTT_CLIENT__PAHO_MQTT_CLIENT_HPP_
+#endif  // VDA5050_CORE__TRANSPORT__PAHO_MQTT_CLIENT_HPP_
