--- vda5050_core/include/vda5050_core/client/order/validation_result.hpp
@@ -1,47 +1,0 @@
-/*
- * Copyright (C) 2025 ROS-Industrial Consortium Asia Pacific
- * Advanced Remanufacturing and Technology Centre
- * A*STAR Research Entities (Co. Registration No. 199702110H)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef VDA5050_CORE__CLIENT__ORDER__VALIDATION_RESULT_HPP_
-#define VDA5050_CORE__CLIENT__ORDER__VALIDATION_RESULT_HPP_
-
-#include <vector>
-
-#include "vda5050_types/error.hpp"
-
-namespace vda5050_core {
-namespace order {
-
-struct ValidationResult
-{
-  /// \brief A vector of error(s) that resulted in an invalid order. Empty if
-  /// order is valid.
-  std::vector<vda5050_types::Error> errors;
-
-  /// \brief Allows use in boolean contexts
-  ///
-  /// \return True if the order is valid, false otherwise.
-  explicit operator bool() const
-  {
-    return errors.empty();
-  }
-};
-
-}  // namespace order
-}  // namespace vda5050_core
-
-#endif  // VDA5050_CORE__CLIENT__ORDER__VALIDATION_RESULT_HPP_
