--- vda5050_core/package.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>vda5050_core</name>
-  <version>0.0.0</version>
-  <description>Common utilities for VDA5050 related workflows</description>
-  <maintainer email="saurabh_kamat@a-star.edu.sg">Saurabh Kamat</maintainer>
-  <license>Apache License 2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <build_depend>vda5050_types</build_depend>
-
-  <test_depend>ament_cmake_cppcheck</test_depend>
-  <test_depend>ament_cmake_cpplint</test_depend>
-  <test_depend>ament_cmake_clang_format</test_depend>
-  <test_depend>ament_cmake_copyright</test_depend>
-  <test_depend>ament_cmake_gmock</test_depend>
-  <test_depend>vda5050_types</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
